Tibu 1.0 Upgrade: 4-Tier System, GIS Auth, and Stability Fixes
- Fixed PHP 8.1 deprecation warnings in inc/functions.php (precision loss in GD functions) that caused "headers already sent" errors during ad creation.
- Fully implemented 4-tier ad system (Free, Premium, VIP, Diamond) with price/duration management and priority sorting.
- Integrated official Google Identity Services (GIS) for secure social login on both sign-in and sign-up pages.
- Added admin toggle for Registration OTP verification and functional "Resend OTP" logic.
- Expanded mobile category grid limit to 16 slots.
- Added Sneakers category and brand sub-categories to automated seeding.
- Verified visual hierarchy including badges, borders, and Diamond shimmer effects.
- Ensured database schema automatically syncs with new tables and settings.

+++ inc/functions.php
@@ -102,8 +102,8 @@ function process_image_upload($file_tmp, $target_dir, $max_width = 800, $user_id
     $filename = md5(uniqid(rand(), true)) . ".jpg";
     $target_file = $target_dir . "/" . $filename;
 
-    $new_width = min($width, $max_width);
-    $new_height = ($height / $width) * $new_width;
+    $new_width = (int)min($width, $max_width);
+    $new_height = (int)(($height / $width) * $new_width);
     $tmp = imagecreatetruecolor($new_width, $new_height);
     imagecopyresampled($tmp, $src, 0, 0, 0, 0, $new_width, $new_height, $width, $height);
     imagejpeg($tmp, $target_file, 80);
@@ -217,38 +217,38 @@ function apply_site_watermark($resource, $seller_info = "") {
     $white = imagecolorallocatealpha($resource, 255, 255, 255, 35); // Slightly lighter for complex text
 
     if (file_exists($font_path) && function_exists('imagettftext')) {
-        $font_size = $width / 14;
+        $font_size = (int)($width / 14);
 
         // Main Centered Site Name
         $bbox = imagettfbbox($font_size, 0, $font_path, $site_text);
         $text_width = $bbox[2] - $bbox[0];
         $text_height = $bbox[7] - $bbox[1];
-        $x = ($width / 2) - ($text_width / 2);
-        $y = ($height / 2) - ($text_height / 2);
+        $x = (int)(($width / 2) - ($text_width / 2));
+        $y = (int)(($height / 2) - ($text_height / 2));
         imagettftext($resource, $font_size, 0, $x, $y, $white, $font_path, $site_text);
 
         // Bottom Right: Site Name + Seller Info (Business name)
         $seller_text = $site_text . ($seller_info ? " | " . $seller_info : "");
-        $small_size = max(8, $width / 45);
+        $small_size = (int)max(8, $width / 45);
         $bbox_small = imagettfbbox($small_size, 0, $font_path, $seller_text);
-        $sx = $width - ($bbox_small[2] - $bbox_small[0]) - 20;
-        $sy = $height - 20;
+        $sx = (int)($width - ($bbox_small[2] - $bbox_small[0]) - 20);
+        $sy = (int)($height - 20);
         imagettftext($resource, $small_size, 0, $sx, $sy, $white, $font_path, $seller_text);
 
         // Tiled secondary watermarks (Jiji Style)
         $tile_text = $site_text;
-        $tile_size = $small_size / 1.5;
+        $tile_size = (int)($small_size / 1.5);
         $tile_color = imagecolorallocatealpha($resource, 255, 255, 255, 15);
         for ($tx = 20; $tx < $width; $tx += ($width/3)) {
             for ($ty = 30; $ty < $height; $ty += ($height/4)) {
-                imagettftext($resource, $tile_size, 45, $tx, $ty, $tile_color, $font_path, $tile_text);
+                imagettftext($resource, $tile_size, 45, (int)$tx, (int)$ty, $tile_color, $font_path, $tile_text);
             }
         }
     } else {
         // Fallback to basic GD font
         $font_size = 5;
-        $x = ($width / 2) - (strlen($site_text) * imagefontwidth($font_size) / 2);
-        $y = ($height / 2) - (imagefontheight($font_size) / 2);
+        $x = (int)(($width / 2) - (strlen($site_text) * imagefontwidth($font_size) / 2));
+        $y = (int)(($height / 2) - (imagefontheight($font_size) / 2));
         imagestring($resource, $font_size, $x, $y, $site_text, $white);
     }
 }
